+++ R/processWithPraat.R
@@ -54,10 +54,10 @@
 #' }
 #'
 #' @param labbcat.url URL to the LaBB-CAT instance
-#' @param matchIds A vector of annotation IDs, e.g. the MatchId column, or the URL column,
+#' @param match.ids A vector of annotation IDs, e.g. the MatchId column, or the URL column,
 #'     of a results set. 
-#' @param startOffsets The start time in seconds, or a vector of start times.
-#' @param endOffsets The end time in seconds, or a vector of end times.
+#' @param start.offsets The start time in seconds, or a vector of start times.
+#' @param end.offsets The end time in seconds, or a vector of end times.
 #' @param praat.script Script to run on each match. This may be a single string or a
 #'     character vector.
 #' @param window.offset In many circumstances, you will want some context before and after
@@ -73,6 +73,8 @@
 #'     gender of the speaker is, including the "participant_gender" attribute will make a
 #'     variable called participant_gender$ available to the praat script, whose value will
 #'     be the gender of the speaker for that segment.
+#' @param no.progress TRUE to supress visual progress bar. Otherwise, progress bar will be
+#'     shown when interactive().
 #' @return A data frame of acoustic measures, one row for each matchId.
 #' 
 #' @seealso \link{praatScriptFormants}
@@ -113,20 +115,21 @@
 #' }
 #' @keywords praat
 #' 
-processWithPraat <- function(labbcat.url, matchIds, startOffsets, endOffsets,
+processWithPraat <- function(labbcat.url, match.ids, start.offsets, end.offsets,
                              praat.script, window.offset=0.0,
-                             gender.attribute="participant_gender", attributes=NULL) {
+                             gender.attribute="participant_gender", attributes=NULL,
+                             no.progress=FALSE) {
 
     ## make the script a single string
     praat.script <- paste(praat.script, collapse="\n")
 
-    ## split matchIds into transcript ID and participant ID
-    transcriptIds <- stringr::str_replace(matchIds, ".*(g_[0-9]+);.*","\\1")
-    participantIds <- stringr::str_replace(matchIds, ".*(p_[0-9]+);.*","\\1")
+    ## split match.ids into transcript ID and participant ID
+    transcriptIds <- stringr::str_replace(match.ids, ".*(g_[0-9]+);.*","\\1")
+    participantIds <- stringr::str_replace(match.ids, ".*(p_[0-9]+);.*","\\1")
 
     ## save CSV file to upload
     upload.file <- tempfile(pattern="processWithPraat.", fileext=".csv")
-    write.table(data.frame(transcriptIds, participantIds, startOffsets, endOffsets),
+    write.table(data.frame(transcriptIds, participantIds, start.offsets, end.offsets),
                 upload.file, sep=",", row.names=FALSE, col.names=TRUE)
     download.file <- tempfile(pattern="praatOutput.", fileext=".csv")
 
@@ -163,7 +166,7 @@ processWithPraat <- function(labbcat.url, matchIds, startOffsets, endOffsets,
     ## wait until the task is finished
     threadId <- resp.json$model$threadId
     pb <- NULL
-    if (interactive()) {
+    if (interactive() && !no.progress) {
         pb <- txtProgressBar(min = 0, max = 100, style = 3)        
     }
     thread <- thread.get(labbcat.url, threadId)
@@ -184,8 +187,9 @@ processWithPraat <- function(labbcat.url, matchIds, startOffsets, endOffsets,
         if (!is.null(thread$percentComplete)) {
             setTxtProgressBar(pb, thread$percentComplete)
         }
+        close(pb)
         if (!is.null(thread$status)) {
-            cat(paste("\n", thread$status, "\n", sep=""))
+            cat(paste(thread$status, "\n", sep=""))
         }
     }
 
@@ -198,7 +202,8 @@ processWithPraat <- function(labbcat.url, matchIds, startOffsets, endOffsets,
     write("\n",file=download.file,append=TRUE)
     
     ## load data into frame
-    results <- read.csv(download.file, header=T)
+    ## (if we don't skip blank lines we get a row too many)
+    results <- read.csv(download.file, header=T, blank.lines.skip=T)
 
     ## tidily remove the downloaded file
     file.remove(download.file)