+++ tests/testthat/test-getMatchingParticipantIds.R
@@ -1,11 +1,11 @@
 labbcat.url <- "https://labbcat.canterbury.ac.nz/demo"
 
-test_that("getMatchingTranscriptIds works", {
+test_that("getMatchingParticipantIds works", {
     skip_on_cran() # don't run tests that depend on external resource on CRAN
     if (!is.null(labbcatCredentials(labbcat.url, "demo", "demo"))) skip("Server not available")
 
-    ids <- getMatchingTranscriptIds(labbcat.url, "/^BR.+/.test(id)")
+    ids <- getMatchingParticipantIds(labbcat.url, "/^BR.+/.test(id)")
     expect_equal(length(ids), 5)
-    expect_true("BR946_RodgerCurragh.eaf" %in% ids)
-    expect_false("QB247_Jacqui.eaf" %in% ids)
+    expect_true("BR946_RodgerCurragh" %in% ids)
+    expect_false("QB247_Jacqui" %in% ids)
 })